refactor(FILE): split the config and metrics god-files by concern
config keeps loading; merge.go, domain.go and print.go take the override
merge, the domain lookup and the obfuscated dump. metrics keeps the gpc
setters; collectors.go holds the declarations, enterprise.go the gpcee set.
